refactor : 에러 페이지 컴포넌트 명 수정 및 디자인 시스템 적용
